requirements.txt
bm3d
FrEIA
imageio
matplotlib
numpy
Pillow
ptwt==0.1.2
scikit-image
scipy
torch==1.11.0
torchvision==0.12.0
wandb
pyradiomics
SimpleITK
clean-fid